refactor(model): remove dead int(size) cast in _convert_text_to_html call
The font_size parameter is unused inside _convert_text_to_html (HTML
applies font-family per char only — actual font-size flows through
_build_insert_css). The int(size) coercion was therefore harmless but
matched the truncation pattern PITFALLS.md warns against. Convert
signature to float to remove the smell entirely.

+++ model/pdf_model.py
@@ -2337,7 +2337,7 @@ class PDFModel:
     def _convert_text_to_html(
         self,
         text: str,
-        font_size: int,
+        font_size: float,
         color: tuple,
         latin_font: str = "helv",
     ) -> str:
@@ -3443,7 +3443,7 @@ class PDFModel:
             )
         else:
             html_content = self._convert_text_to_html(
-                new_text, int(size), color, latin_font=resolve_result.resolved_font
+                new_text, size, color, latin_font=resolve_result.resolved_font
             )
 
         # Preserve the original visual line height to keep committed text
